requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
backcall==0.1.0           # via ipython
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
decorator==4.4.2          # via ipython, traitlets
django-environ==0.4.5     # via -r requirements.in
django-extensions==2.2.9  # via -r requirements.in
django-filter==2.3.0      # via -r requirements.in
django==2.2.13            # via -r requirements.in, django-filter, djangorestframework, djangorestframework-simplejwt, drf-yasg
djangorestframework-simplejwt==4.4.0  # via -r requirements.in
djangorestframework==3.11.0  # via -r requirements.in, djangorestframework-simplejwt, drf-yasg
drf-yasg==1.17.1          # via -r requirements.in
factory-boy==2.12.0       # via -r requirements.in
faker==4.1.0              # via factory-boy
flake8-print==3.1.4       # via -r requirements.in
flake8==3.8.2             # via flake8-print
gunicorn==20.0.4          # via -r requirements.in
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via flake8
inflection==0.4.0         # via drf-yasg
ipdb==0.13.2              # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.15.0           # via -r requirements.in, ipdb
itypes==1.2.0             # via coreapi
jedi==0.17.0              # via ipython
jinja2==2.11.2            # via coreschema
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
packaging==20.4           # via drf-yasg
parso==0.7.0              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.2.0          # via -r requirements.in
prompt-toolkit==3.0.5     # via ipython
psycopg2-binary==2.8.5    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pycodestyle==2.6.0        # via flake8, flake8-print
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via ipython
pyjwt==1.7.1              # via djangorestframework-simplejwt
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via faker
pytz==2020.1              # via django
requests==2.23.0          # via -r requirements.in, coreapi
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via drf-yasg
six==1.15.0               # via django-extensions, drf-yasg, flake8-print, packaging, pip-tools, python-dateutil, traitlets
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
traitlets==4.3.3          # via ipython
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.9           # via requests
wcwidth==0.2.3            # via prompt-toolkit
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
